FAFB13 brain and neuropil surface objects

FAFB13.surf describes the neuropil surface of the FAFB13 brain. FAFB13NP.surf is a surface object for the Ito et al (2014) neuropil regions in FAFB13 space

Details

These neuropil regions are generated from the JFRC2.surf and JFRC2NP.surf objects, respectively. Because they are not defined from FAFB13 itself but transformed using bridging registrations, they are only as good as those bridging registrations. In particular although the elm.landmarks used to bridge FAFB13-JFRC2013 are well defined in the fly's right brain hemisphere, they are somewhat less dense in other areas. Therefore these surfaces are clearly likely less accurate in the left brain hemisphere and show significanct displacements in the optic lobes. Note that updated landmarks in March 2017 significantly improved the situation.

The acronyms for the neuropil regions are those used in Ito et al. 2014 and can also be looked up at http://virtualflybrain.org.

See also

FAFB13, FAFB13

Examples

boundingbox(FAFB13.surf)

#>          [,1]     [,2]     [,3]
#> [1,] 286193.4 156861.7    726.1
#> [2,] 913545.7 467546.8 235903.3
#> attr(,"class")
#> [1] "boundingbox"

FAFB13NP.surf$RegionList

#>  [1] "AME_R"    "LO_R"     "NO"       "BU_R"     "PB"       "LH_R"    
#>  [7] "LAL_R"    "SAD"      "CAN_R"    "AMMC_R"   "ICL_R"    "VES_R"   
#> [13] "IB_R"     "ATL_R"    "CRE_R"    "MB_PED_R" "MB_VL_R"  "MB_ML_R" 
#> [19] "FLA_R"    "LOP_R"    "EB"       "AL_R"     "ME_R"     "FB"      
#> [25] "SLP_R"    "SIP_R"    "SMP_R"    "AVLP_R"   "PVLP_R"   "IVLP_R"  
#> [31] "PLP_R"    "AOTU_R"   "GOR_R"    "MB_CA_R"  "SPS_R"    "IPS_R"   
#> [37] "SCL_R"    "EPA_R"    "GNG"      "PRW"      "GA_R"     "AME_L"   
#> [43] "LO_L"     "BU_L"     "LH_L"     "LAL_L"    "CAN_L"    "AMMC_L"  
#> [49] "ICL_L"    "VES_L"    "IB_L"     "ATL_L"    "CRE_L"    "MB_PED_L"
#> [55] "MB_VL_L"  "MB_ML_L"  "FLA_L"    "LOP_L"    "AL_L"     "ME_L"    
#> [61] "SLP_L"    "SIP_L"    "SMP_L"    "AVLP_L"   "PVLP_L"   "IVLP_L"  
#> [67] "PLP_L"    "AOTU_L"   "GOR_L"    "MB_CA_L"  "SPS_L"    "IPS_L"   
#> [73] "SCL_L"    "EPA_L"    "GA_L"    

not_run({
  plot3d(FAFB13.surf)
  # a short cut which plots the surface in semi-transparent grey.
  clear3d()
  plot3d(FAFB13)

  # plot right lateral horn
  plot3d(FAFB13NP.surf, "LH_R", alpha=0.4)
  # plot both antennal lobe surfaces (second arg is regex if no exact match)
  # note use of initial anchor (^) to avoid matches to LAL
  plot3d(FAFB13NP.surf, "^AL", alpha=0.4)
})