ELM Landmarks created by Bock & Saalfeld groups at Janelia

These are used to map locations in EM (FAFB) space to JFRC2013 (aka DPX) space, in conjunction with a thin plate spline library.

Details

X,Y,Z correspond to locations in JFRC2013 template brain. However, they are in raw voxel coordinates. In contrast, X1,Y1,Z1 (which contain locations in FAFB space) are in nm.

Examples

head(elm.landmarks)

#>   Label  Use        X        Y        Z       X1       Y1     Z1
#> 1  Pt-0 TRUE 704.8786 264.6964 210.1924 460676.6 221203.6  95445
#> 2  Pt-1 TRUE 730.4794 264.7117 211.8637 471134.9 225319.8  92015
#> 3  Pt-2 TRUE 736.0885 305.8448 221.2121 470906.8 235640.1  97265
#> 4  Pt-3 TRUE 702.6678 303.7415 224.4841 459799.2 232858.2 100870
#> 5  Pt-4 TRUE 714.8208 204.4144 211.1626 466386.2 203906.7  79380
#> 6  Pt-5 TRUE 600.5362 288.7840 382.4726 431200.3 195286.6 165970
#> 

## Not run: ------------------------------------
# # Show landmarks in the context of the JFRC2013 template brain
# library(nat.flybrains)
# plot3d(JFRC2013)
# # note that the landmarks are in raw voxel coordinates and must be scaled
# xyz=scale(elm.landmarks[,c("X","Y","Z")], scale = 1/voxdims(JFRC2013), center = FALSE)
# spheres3d(xyz, col = ifelse(elm.landmarks$Use, "green", 'red'), radius = 4)
## ---------------------------------------------