Functions that have been removed or moved to other packages

Functions that have been removed or moved to other packages

moved

distal_to has been moved to the nat package.

simplify_neuron has been moved to the nat package.

stitch_neuron has been moved to the nat package.

stitch_neurons has been moved to the nat package.

prune_twigs has been moved to the nat package.

tpsreg has been moved to the nat package.

plot3d.ngraph has been moved to the nat package.