Convert coordinates between different FAFB assemblies

fafb_world_mapper provides a low-level interface to map coordinates based on the tem-services API available from inside the HHMI Janelia VPN. The hope is to make this service generally available in due course.

End users are not presently expected to use this directly. Instead they should use xform_brain to transform data. See examples.

fafb_world_mapper(
  xyz,
  from,
  to,
  baseurl = paste0("http://tem-services.int.janelia.org:8080/render-ws/v1/",
    "owner/flyTEM/project/FAFB00/stack"),
  method = c("single", "many"),
  ...
)

Arguments

xyz
xyz coordinates (in nm)

from, to
Assemblies to convert between e.g. "v12", "v13", "v14".

baseurl
path to Janelia tem-services server (not externally visible at the moment)

method
Whether to try and convert multiple coordinates at once. Defaults to 'single'. The 'many' method currently appears fragile and not much faster.

...
Additional arguments (currently unused)

Examples

if (FALSE) {
FAFB12.surf=xform_brain(JFRC2013.surf, sample = JFRC2013, reference = FAFB12)
FAFB13.surf=xform_brain(JFRC2013.surf, sample=JFRC2013, ref=FAFB13)
# current assembly
FAFB.surf=xform_brain(JFRC2013.surf, sample=JFRC2013, ref=FAFB)
}