Fetch neurons from FAFB CATMAID server, transforming them as appropriate

fetchn_fafb is a thin wrapper around the catmaid::read.neurons.catmaid function with the additional convenience of automating conversion to a reference brain. Input can also be a pre-downloaded neuronlist.

fetchdp_fafb extends fetchn_fafb by additionally converting neurons to the dotprops representation suitable for nblast.

fetchn_fafb(skids, mirror = TRUE, conn = NULL,
  reference = nat.flybrains::FCWB)

fetchdp_fafb(skids, mirror = TRUE, conn = NULL, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

mirror
whether to mirror the neuron (default TRUE since FlyCircuit neurons are on fly's left and most FAFB tracings are on fly's right.)

conn
a catmaid connection object (see catmaid_connection)

reference
The reference brain to which the neurons will be transformed, either a templatebrain object such as FCWB or a character vector naming such an object.

...
Additional parameters passed to nblast

Examples

# NOT RUN {
# fetch all neurons with an annotation exactly matching "KC" and convert
# to JFRC2013 space
kcs=fetchn_fafb("annotation:^KC$", reference=JFRC2013)

# supplying a neuronlist, it will mirror and convert the neurons
# to a reference brain
dc_nl=fetchn_fafb(dense_core_neurons, reference=JFRC2013)
# }