Open FAFB CATMAID in browser at a given XYZ location

Open FAFB CATMAID in browser at a given XYZ location

open_fafb(x, s = rgl::select3d(), mirror = FALSE, sample = elmr::FAFB,
  zoom = 1, open = interactive(), active_skeleton_id = NULL,
  active_node_id = NULL, ...)

Arguments

x
A numeric vector or any object compatible with xyzmatrix (see details)

s
A selection function of the type returned by select3d

mirror
Whether to mirror the point to the opposite side of the brain

sample
The template brain space associated with the coordinates in x

zoom
The CATMAID zoom factor (defaults to 1)

open
Whether to open the url in the browser or simply return it. Defaults to TRUE when R is running in interactive mode.

active_skeleton_id, active_node_id
Set highlighted skeleton and node in CATMAID.

...
Additional arguments to be added to URL.

Details

Note that if object x contains exactly one point then CATMAID will be opened immediately at that location, whereas if there is more than 1 point, the function will stop and wait for the user to make an interactive selection in a rgl window.

As a special case, if x is a data.frame generated by catmaid_get_treenodes_detail then URLs will be generated for every row (i.e. each treenode). See examples.

See also

Examples

open_fafb(c(316, 143, 26), sample=JFRC2013, open=FALSE)

#> [1] "https://neuropil.janelia.org/tracing/fafb/v14/?pid=1&zp=61258&yp=291205&xp=580355&tool=tracingtool&sid0=5&s0=1.000000"

library(nat)
# NOT RUN {
open3d()
plot3d(kcs20)
# waits for used to draw a selection rectangle
open_fafb(kcs20, sample=FCWB)
# same but mirrors selected points to opposite hemisphere
open_fafb(kcs20, sample=FCWB, mirror=TRUE)
# }
# fetch a neuron and extract the node ids for a given tag
n=read.neuron.catmaid(16)
tagged = n$tags[['TODO']]
tagged_details <- catmaid_get_treenodes_detail(tagged)
# now find FAFB URLs
tagged_details$url <- open_fafb(tagged_details)
# take a look at results
head(tagged_details)

#>   treenode_id parent_id      x      y      z confidence radius skid
#> 1    10362518  10362517 363719 149281 181280          5     -1   16
#> 2    18699746  18699745 455425 245487  33680          5     -1   16
#> 3    18699757  18699755 455145 244639  34680          5     -1   16
#> 4    18699947  18699945 467610 246768  30480          5     -1   16
#> 5    18700002  18700000 465554 248064  28680          5     -1   16
#> 6    18700064  18700063 460012 232579  26840          5     -1   16
#>   edition_time user_id
#> 1   1490896453     117
#> 2   1508753021     167
#> 3   1508753040     167
#> 4   1508753591     167
#> 5   1508753660     167
#> 6   1508753889     167
#>                                                                                                                                                                    url
#> 1 https://neuropil.janelia.org/tracing/fafb/v14/?pid=1&zp=181280&yp=149281&xp=363719&tool=tracingtool&sid0=5&s0=1.000000&active_skeleton_id=16&active_node_id=10362518
#> 2  https://neuropil.janelia.org/tracing/fafb/v14/?pid=1&zp=33680&yp=245487&xp=455425&tool=tracingtool&sid0=5&s0=1.000000&active_skeleton_id=16&active_node_id=18699746
#> 3  https://neuropil.janelia.org/tracing/fafb/v14/?pid=1&zp=34680&yp=244639&xp=455145&tool=tracingtool&sid0=5&s0=1.000000&active_skeleton_id=16&active_node_id=18699757
#> 4  https://neuropil.janelia.org/tracing/fafb/v14/?pid=1&zp=30480&yp=246768&xp=467610&tool=tracingtool&sid0=5&s0=1.000000&active_skeleton_id=16&active_node_id=18699947
#> 5  https://neuropil.janelia.org/tracing/fafb/v14/?pid=1&zp=28680&yp=248064&xp=465554&tool=tracingtool&sid0=5&s0=1.000000&active_skeleton_id=16&active_node_id=18700002
#> 6  https://neuropil.janelia.org/tracing/fafb/v14/?pid=1&zp=26840&yp=232579&xp=460012&tool=tracingtool&sid0=5&s0=1.000000&active_skeleton_id=16&active_node_id=18700064
# NOT RUN {
# copy results to clipboard e.g. to paste into a spreadsheet
library(clipr)
write_clip(tagged_details)
# }