Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids

catmaid skeleton ids (see catmaid_skids)

type

Whether to summarise contributinons towards arbour (nodes) or synapses

...

Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x

A data frame containing the output of summarise_contribution

auth

The minimum percentage of the total effort for authorship

ackn

The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   12 110422          Jacob Ratliff 15.7686920 15.76869   auth
#> 2   11  80683        bailey harrison 11.5218469 27.29054   auth
#> 3   26  70493             Jawaid Ali 10.0666751 37.35721   auth
#> 4   13  63207           Addy Adesina  9.0262059 46.38342   auth
#> 5   54  42341        Ben Koppenhaver  6.0464598 52.42988   auth
#> 6   25  39614          Lucia Kmecova  5.6570336 58.08691   auth
#> 7   53  37952      Najla Masoodpanah  5.4196935 63.50661   auth
#> 8    4  34366           Corey Fisher  4.9075987 68.41421    ack
#> 9   61  31177         Benjamin Gorko  4.4521971 72.86640    ack
#> 10  17  28372              Adam John  4.0516322 76.91803    ack
#> 11 117  27320         Ruairi Roberts  3.9014025 80.81944    ack
#> 12  22  21093     Michael Lingelbach  3.0121626 83.83160    ack
#> 13  50  18611       Shahrozia Imtiaz  2.6577233 86.48932    ack
#> 14  29  15865         Nadiya Sharifi  2.2655838 88.75491    ack
#> 15  75  14828          Philipp Ranft  2.1174962 90.87240    ack
#> 16  55   9288          Greg Jefferis  1.3263626 92.19877    ack
#> 17  27   7852             Sri Murthy  1.1212962 93.32006    ack
#> 18 120   4889 Amelia Edmondson-Stait  0.6981683 94.01823    ack
#> 19 118   4260         Clement Hallou  0.6083446 94.62657    ack
#> 20  94   4053       Philipp Schlegel  0.5787842 95.20536    ack
#> 21  28   3769           Bindu Gampah  0.5382279 95.74359    ack
#> 22 123   3144             Fiona Love  0.4489755 96.19256    ack
#> 23  76   3057           Steven Calle  0.4365515 96.62911    ack
#> 24   2   3039              Davi Bock  0.4339810 97.06310    ack

write_authors(pnsc, auth=3.0)

#> Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Ruairi Roberts, Michael Lingelbach

write_ack(pnsc)

#> We thank Corey Fisher, Benjamin Gorko, Adam John, Ruairi Roberts, Michael Lingelbach, Shahrozia Imtiaz, Nadiya Sharifi, Philipp Ranft, Greg Jefferis, Sri Murthy, Amelia Edmondson-Stait, Clement Hallou, Philipp Schlegel, Bindu Gampah, Fiona Love, Steven Calle, Davi Bock for contributing 33.6 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')