Thin plate spline registrations via nat::xform and friends
Source: R/xform_tps.R

tpsreg creates an object encapsulating a thin plate spine transform mapping a paired landmark set.

xformpoints.tpsreg enables xform and friends to transform 3d vertices (or more complex objects containing 3d vertices) using a thin plate spline mapping stored in a tpsreg object.

tpsreg(sample, reference, ...)

# S3 method for tpsreg
xformpoints(reg, points, swap = NULL, ...)

Arguments

sample, reference
Matrices defining the sample (or floating) and reference (desired target after transformation) spaces. See details.

...
additional arguments passed to xformpoints.tpsreg

reg
The tpsreg registration object

points
The 3D points to transform

swap
Whether to change the direction of registration (default of NULL checks if reg has a attr('swap'=TRUE)) otherwise

Details

Note that we use the nat convention for naming the sample/reference space arguments but these actually clash with the nomenclature in the underlying Morpho::tps3d function.

refmat (Morpho3d) == sample (nat)

tarmat (Morpho3d) == reference (nat)

See also

reglist, read.landmarks

Examples

## A full worked example of using landmarks based registration to construct
## a mirroring registration from one side of the brain to the other.

# read in set of landmarks defined in FAFB CATMAID
emlandmarks=read.neurons.catmaid('annotation:^GJLandmark')
# Match up L and R pairs
library(stringr)
emlandmarks[,'side']=str_match(emlandmarks[,'name'], "([LR]) Landmark")[,2]
emlandmarks[,'shortname']=str_match(emlandmarks[,'name'], "(.*)([LR]) Landmark.*")[,2]
emlandmarks[,'shortname']=sub("[_ ]+$", "", emlandmarks[,'shortname'])
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:nat’:
#> 
#>     intersect, setdiff, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

lmpairs=inner_join(
  filter(emlandmarks[,], side=="L"),
  filter(emlandmarks[,], side=="R"),
  by='shortname', suffix=c(".L",".R"))

# find mean xyz position of each landmark (they are drawn as a little cross)
lmxyz=t(sapply(emlandmarks, function(x) colMeans(xyzmatrix(x))))
# construct thin plate splines registration
mirror_reg=tpsreg(
  lmxyz[as.character(lmpairs$skid.R),],
  lmxyz[as.character(lmpairs$skid.L),]
)
# map RHS DA2 PNs onto left and compare with LHS neurons
da2pns.R=read.neurons.catmaid('name:^PN Glomerulus DA2')
da2pns.L=read.neurons.catmaid('name:Left uPN mALT DA2')

#> Warning: No matches for query "name:Left uPN mALT DA2"!

#> Error in names(res) <- `*vtmp*`: attempt to set an attribute on NULL

da2pns.R.L=xform(da2pns.R, reg = mirror_reg)
plot(da2pns.L, col='red')

#> Error in plot(da2pns.L, col = "red"): object 'da2pns.L' not found

plot(da2pns.R.L, col='blue', add=T)

#> Error in plot.xy(xy.coords(x, y), type = type, ...): plot.new has not been called yet