templatebrain for assembled FAFB full adult female brain EM volume

FAFB templatebrain objects encapsulate the physical dimensions of assembled FAFB serial section data. There is a different templatebrain for each assembly version.They are used to define the transformations between FAFB and other templates like JFRC2013 (see xform_brain)

FAFB12 Assembly v12 of FAFB full adult female brain EM volume (2016-04-01)

FAFB11 Assembly v11 of FAFB full adult female brain EM volume (2016-02-09)

Details

FAFB12 is the default FAFB assembly since 2016-04-01.

See also

templatebrain, xform_brain

Examples

FAFB12

#> === Template Brain === 
#> Name: FAFB12 
#> Short Name: FAFB12 
#> Type: Stitched serial section EM volume of whole adult female brain 
#> Sex:  F 
#> Voxel size:
#>   x = 4 nm
#>   y = 4 NA
#>   z = 35 NA
#> Bounding box (nm):
#>   x = 0, y = 0, z = 0,
#>   x = 919160, y = 571552, z = 247170.
#> Description: 
#> DOI:
#> 

dim(FAFB12)

#> [1] 229791 142889   7063
#> 

voxdims(FAFB12)

#> [1]  4  4 35
#> 

boundingbox(FAFB12)

#>        [,1]   [,2]   [,3]
#> [1,]      0      0      0
#> [2,] 919160 571552 247170
#> attr(,"class")
#> [1] "boundingbox"
#>