templatebrain for assembled FAFB full adult female brain EM volume

FAFB Current default assembly of FAFB (an alias for FAFB13 since 2016-12-07)

FAFB14 Assembly v14 of FAFB full adult female brain EM volume (2017-10-07 - ).

FAFB13 Assembly v13 of FAFB full adult female brain EM volume (2016-12-07 - 2017-10-06).

FAFB12 Assembly v12 of FAFB full adult female brain EM volume (2016-04-01 to 2016-12-06)

FAFB11 Assembly v11 of FAFB full adult female brain EM volume (2016-02-09)

Details

FAFB templatebrain objects encapsulate the physical dimensions of assembled FAFB serial section data. There is a different templatebrain for each assembly version. They are used to define the transformations between FAFB and other templates like JFRC2013 (see xform_brain)

FAFB14 is the default FAFB assembly since 2017-10-07. If you want to use a specific assembly then refer to it with a versioned name. If you just want to use the current working assembly, then use FAFB.

See also

Examples

FAFB13

#> === Template Brain === 
#> Name: FAFB13 
#> Short Name: FAFB13 
#> Type: Stitched serial section EM volume of whole adult female brain 
#> Sex:  F 
#> Dimensions:229791 x 142889 x 7063 voxels
#> Voxel size:
#>   x = 4 nm
#>   y = 4 NA
#>   z = 35 NA
#> Bounding box (nm):
#>   x = 0, y = 0, z = 0,
#>   x = 919160, y = 571552, z = 247170.
#> Description: 
#> DOI: https://doi.org/10.1101/140905

dim(FAFB13)

#> [1] 229791 142889   7063

voxdims(FAFB13)

#> [1]  4  4 35

boundingbox(FAFB13)

#>        [,1]   [,2]   [,3]
#> [1,]      0      0      0
#> [2,] 919160 571552 247170
#> attr(,"class")
#> [1] "boundingbox"