ELM Landmarks created by Bock & Saalfeld groups at Janelia

These are used to map locations in EM (FAFB) space to JFRC2013 (aka DPX) space, in conjunction with a thin plate spline library.

elm.landmarks.12 contains the landmark positions used for the FAFB12 assembly as a historical record.

Details

X,Y,Z correspond to locations in JFRC2013 template brain. However, they are in raw voxel coordinates. In contrast, X1,Y1,Z1 (which contain locations in FAFB space) are in nm.

The current iteration of these landmarks are in FAFB13 space

References

See https://github.com/saalfeldlab/elm for the original source of the landmarks.

Examples

head(elm.landmarks)

#>   Label  Use        X         Y         Z       X1       Y1       Z1
#> 1  Pt-0 TRUE 702.1487 285.40275 209.73076 591360.1 287783.6 97510.11
#> 2  Pt-1 TRUE 571.4001  38.85996 287.05954 525666.5 172470.4 80994.73
#> 3  Pt-2 TRUE 715.8113 213.29936 217.39349 595391.6 263523.1 84156.77
#> 4  Pt-3 TRUE 513.0022 198.00197 217.79409 501080.2 253465.3 98233.34
#> 5  Pt-5 TRUE 602.0898 271.07851  17.30991 528543.2 338099.9 19349.01
#> 6  Pt-6 TRUE 867.0125  31.91925 276.22344 670999.9 179097.9 67561.69

## Not run: ------------------------------------
# # Show landmarks in the context of the JFRC2013 template brain
# library(nat.flybrains)
# plot3d(JFRC2013)
# # note that the landmarks are in raw voxel coordinates and must be scaled
# xyz=scale(elm.landmarks[,c("X","Y","Z")], scale = 1/voxdims(JFRC2013), center = FALSE)
# spheres3d(xyz, col = ifelse(elm.landmarks$Use, "green", 'red'), radius = 4)
## ---------------------------------------------