NBLAST EM tracing against flycircuit (or other databases of) neurons

NBLAST EM tracing against flycircuit (or other databases of) neurons

nblast_fafb(skids, db = NULL, conn = NULL, mirror = TRUE,
  normalised = TRUE, reverse = FALSE, reference = nat.flybrains::FCWB,
  .parallel = TRUE, ...)

Arguments

skids

catmaid skeleton ids (see catmaid_skids)

db

A neuronlist object containing neurons to search. Defaults to the value of options('nat.default.neuronlist').

conn

a catmaid connection object (see catmaid_connection)

mirror

whether to mirror the neuron (default TRUE since flycircuit neurons are on fly's left and most FAFB tracings are on fly's right.)

normalised

Whether to return normalised NBLAST scores

reverse

Treat the FAFB skeleton as NBLAST target rather than query (sensible if db contains partial skeletons/tracts; default FALSE).

reference

The reference brain to which the neurons will be transformed, either a templatebrain object such as FCWB or a character vector naming such an object.

.parallel

Whether to parallelise the nblast search (see details of nblast for how to use the parallel interface provided by the doMC package.)

...

Additional parameters passed to nblast

Value

an object of class nblastfafb for which plot3d, summary and hist methods exist.

Details

Still depends on having a neuronlist containing registered neurons (usually from flycircuit.tw). The example code downloads a set of projection neurons. The full data must be requested from Greg Jefferis.

See also

nblast

Examples

## Not run: ------------------------------------
# # first load neuronlist object containing registered neurons (see details)
# allpndps=flycircuit::load_si_data('allpndps.rds')
# # ... and set that as the default for queries and plotting
# options(nat.default.neuronlist='allpndps')
# 
# # then make sure you are loged in to catmaid server
# # catmaid::catmaid_login(<your connection args>)
# 
# # nblast neuron 27884
# PN27884f=nblast_fafb(27884, mirror = FALSE)
# # summary table of results
# summary(PN27884f)
# # plot results, just top hit
# plot3d(PN27884f, hits=1)
## ---------------------------------------------