Open FAFB CATMAID in browser at a given XYZ location

Open FAFB CATMAID in browser at a given XYZ location

open_fafb(x, s = rgl::select3d(), mirror = FALSE, sample = elmr::FAFB13,
  zoom = 1, open = interactive(), active_skeleton_id = NULL,
  active_node_id = NULL, ...)

Arguments

x
A numeric vector or any object compatible with xyzmatrix (see details)

s
A selection function of the type returned by select3d

mirror
Whether to mirror the point to the opposite side of the brain

sample
The template brain space associated with the coordinates in x

zoom
The CATMAID zoom factor (defaults to 1)

open
Whether to open the url in the browser or simply return it. Defaults to TRUE when R is running in interactive mode.

active_skeleton_id, active_node_id
Set highlighted skeleton and node in CATMAID.

...
Additional arguments to be added to URL.

Details

Note that if object x contains exactly one point then CATMAID will be opened immediately at that location, whereas if there is more than 1 point, the function will stop and wait for the user to make an interactive selection in a rgl window.

See also

xform_brain

Examples

open_fafb(c(316, 143, 26), sample=JFRC2013, open=FALSE)

#> [1] "https://neuropil.janelia.org/tracing/fafb/v13/?pid=1&zp=53617&yp=370558&xp=653365&tool=tracingtool&sid0=5&s0=1.000000"

library(nat)
not_run({
  open3d()
  plot3d(kcs20)
  # waits for used to draw a selection rectangle
  open_fafb(kcs20, sample=FCWB)
  # same but mirrors selected points to opposite hemisphere
  open_fafb(kcs20, sample=FCWB, mirror=TRUE)
})