Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria. suggest_authorship suggests how to acknowledge individuals write_authors produces a list of authors write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributinons towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   12 110309          Jacob Ratliff 13.1165182 13.11652   auth
#> 2  117 108854         Ruairi Roberts 12.9435084 26.06003   auth
#> 3   11  80570        bailey harrison  9.5803413 35.64037   auth
#> 4   26  73369             Jawaid Ali  8.7240916 44.36446   auth
#> 5   13  63159           Addy Adesina  7.5100506 51.87451   auth
#> 6   54  42990        Ben Koppenhaver  5.1118142 56.98632   auth
#> 7   25  41753          Lucia Kmecova  4.9647262 61.95105    ack
#> 8   53  39790      Najla Masoodpanah  4.7313117 66.68236    ack
#> 9    4  34884           Corey Fisher  4.1479537 70.83032    ack
#> 10  61  33473         Benjamin Gorko  3.9801758 74.81049    ack
#> 11  17  28462              Adam John  3.3843326 78.19482    ack
#> 12  22  20918     Michael Lingelbach  2.4872978 80.68212    ack
#> 13  50  18732       Shahrozia Imtiaz  2.2273669 82.90949    ack
#> 14  94  17904       Philipp Schlegel  2.1289119 85.03840    ack
#> 15  29  15998         Nadiya Sharifi  1.9022750 86.94068    ack
#> 16 109  15889             Alex Bates  1.8893142 88.82999    ack
#> 17  75  14822          Philipp Ranft  1.7624404 90.59243    ack
#> 18  55   9738          Greg Jefferis  1.1579169 91.75035    ack
#> 19  27   7886             Sri Murthy  0.9377010 92.68805    ack
#> 20 127   5193             Lisa Marin  0.6174843 93.30553    ack
#> 21 167   5179           Arian Jamasb  0.6158196 93.92135    ack
#> 22 164   4691    Mahmoud Elbahnasawi  0.5577930 94.47915    ack
#> 23  44   4617             Ala Haddad  0.5489939 95.02814    ack
#> 24 120   4613 Amelia Edmondson-Stait  0.5485182 95.57666    ack
#> 25 118   4262         Clement Hallou  0.5067819 96.08344    ack
#> 26 123   3923             Fiona Love  0.4664724 96.54991    ack
#> 27  28   3769           Bindu Gampah  0.4481607 96.99807    ack
#> 28  76   3422           Steven Calle  0.4068999 97.40497    ack
#> 29   2   3039              Davi Bock  0.3613585 97.76633    ack

write_authors(pnsc, auth=3.0)

#> Jacob Ratliff, Ruairi Roberts, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Shahrozia Imtiaz, Philipp Schlegel, Nadiya Sharifi, Alex Bates, Philipp Ranft, Greg Jefferis, Sri Murthy, Lisa Marin, Arian Jamasb, Mahmoud Elbahnasawi, Ala Haddad, Amelia Edmondson-Stait, Clement Hallou, Fiona Love, Bindu Gampah, Steven Calle, Davi Bock for contributing 40.8 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')