Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids

catmaid skeleton ids (see catmaid_skids)

type

Whether to summarise contributinons towards arbour (nodes) or synapses

...

Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x

A data frame containing the output of summarise_contribution

auth

The minimum percentage of the total effort for authorship

ackn

The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   12 110422          Jacob Ratliff 15.7902037 15.79020   auth
#> 2   11  80683        bailey harrison 11.5375650 27.32777   auth
#> 3   26  70398             Jawaid Ali 10.0668233 37.39459   auth
#> 4   13  63207           Addy Adesina  9.0385196 46.43311   auth
#> 5   54  42322        Ben Koppenhaver  6.0519915 52.48510   auth
#> 6   25  39600          Lucia Kmecova  5.6627490 58.14785   auth
#> 7   53  37918      Najla Masoodpanah  5.4222251 63.57008   auth
#> 8    4  34354           Corey Fisher  4.9125777 68.48265    ack
#> 9   61  31142         Benjamin Gorko  4.4532659 72.93592    ack
#> 10  17  28372              Adam John  4.0571594 76.99308    ack
#> 11 117  26607         Ruairi Roberts  3.8047667 80.79785    ack
#> 12  22  21093     Michael Lingelbach  3.0162718 83.81412    ack
#> 13  50  18610       Shahrozia Imtiaz  2.6612060 86.47532    ack
#> 14  29  15861         Nadiya Sharifi  2.2681026 88.74343    ack
#> 15  75  14827          Philipp Ranft  2.1202419 90.86367    ack
#> 16  55   9288          Greg Jefferis  1.3281720 92.19184    ack
#> 17  27   7848             Sri Murthy  1.1222539 93.31410    ack
#> 18 120   4888 Amelia Edmondson-Stait  0.6989777 94.01307    ack
#> 19 118   4260         Clement Hallou  0.6091745 94.62225    ack
#> 20  94   4044       Philipp Schlegel  0.5782868 95.20053    ack
#> 21  28   3769           Bindu Gampah  0.5389621 95.73950    ack
#> 22 123   3143             Fiona Love  0.4494450 96.18894    ack
#> 23  76   3056           Steven Calle  0.4370041 96.62595    ack
#> 24   2   3039              Davi Bock  0.4345731 97.06052    ack

write_authors(pnsc, auth=3.0)

#> Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Ruairi Roberts, Michael Lingelbach

write_ack(pnsc)

#> We thank Corey Fisher, Benjamin Gorko, Adam John, Ruairi Roberts, Michael Lingelbach, Shahrozia Imtiaz, Nadiya Sharifi, Philipp Ranft, Greg Jefferis, Sri Murthy, Amelia Edmondson-Stait, Clement Hallou, Philipp Schlegel, Bindu Gampah, Fiona Love, Steven Calle, Davi Bock for contributing 33.5 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')