Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria. suggest_authorship suggests how to acknowledge individuals write_authors produces a list of authors write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributinons towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1  117 146142         Ruairi Roberts 15.2216187 15.22162   auth
#> 2   12 110324          Jacob Ratliff 11.4909462 26.71256   auth
#> 3   11  80447        bailey harrison  8.3790667 35.09163   auth
#> 4   26  74529             Jawaid Ali  7.7626693 42.85430   auth
#> 5   13  62988           Addy Adesina  6.5606008 49.41490   auth
#> 6   94  49277       Philipp Schlegel  5.1325129 54.54741   auth
#> 7   54  42981        Ben Koppenhaver  4.4767445 59.02416    ack
#> 8   25  41756          Lucia Kmecova  4.3491529 63.37331    ack
#> 9   53  39791      Najla Masoodpanah  4.1444857 67.51780    ack
#> 10   4  34968           Corey Fisher  3.6421396 71.15994    ack
#> 11  61  33676         Benjamin Gorko  3.5075696 74.66751    ack
#> 12  17  30539              Adam John  3.1808311 77.84834    ack
#> 13  22  23722     Michael Lingelbach  2.4707972 80.31914    ack
#> 14 109  20945             Alex Bates  2.1815550 82.50069    ack
#> 15  50  18729       Shahrozia Imtiaz  1.9507445 84.45143    ack
#> 16 164  16947    Mahmoud Elbahnasawi  1.7651378 86.21657    ack
#> 17 167  16313           Arian Jamasb  1.6991027 87.91568    ack
#> 18  29  16121         Nadiya Sharifi  1.6791047 89.59478    ack
#> 19  75  14785          Philipp Ranft  1.5399518 91.13473    ack
#> 20 127  10843             Lisa Marin  1.1293674 92.26410    ack
#> 21  55  10190          Greg Jefferis  1.0613533 93.32545    ack
#> 22  27   7885             Sri Murthy  0.8212729 94.14673    ack
#> 23 123   6551             Fiona Love  0.6823283 94.82905    ack
#> 24 118   5064         Clement Hallou  0.5274478 95.35650    ack
#> 25 120   4582 Amelia Edmondson-Stait  0.4772444 95.83375    ack
#> 26  44   4582             Ala Haddad  0.4772444 96.31099    ack
#> 27  28   3769           Bindu Gampah  0.3925653 96.70356    ack
#> 28  76   3605           Steven Calle  0.3754837 97.07904    ack
#> 29   2   3040              Davi Bock  0.3166353 97.39567    ack

write_authors(pnsc, auth=3.0)

#> Ruairi Roberts, Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Philipp Schlegel, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Alex Bates, Shahrozia Imtiaz, Mahmoud Elbahnasawi, Arian Jamasb, Nadiya Sharifi, Philipp Ranft, Lisa Marin, Greg Jefferis, Sri Murthy, Fiona Love, Clement Hallou, Amelia Edmondson-Stait, Ala Haddad, Bindu Gampah, Steven Calle, Davi Bock for contributing 42.8 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')