Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria. suggest_authorship suggests how to acknowledge individuals write_authors produces a list of authors write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributinons towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1  117 137181         Ruairi Roberts 14.5301001 14.53010   auth
#> 2   12 110329          Jacob Ratliff 11.6859581 26.21606   auth
#> 3   11  80447        bailey harrison  8.5208809 34.73694   auth
#> 4   26  74439             Jawaid Ali  7.8845184 42.62146   auth
#> 5   13  62989           Addy Adesina  6.6717437 49.29320   auth
#> 6   94  49229       Philipp Schlegel  5.2142957 54.50750   auth
#> 7   54  42986        Ben Koppenhaver  4.5530422 59.06054    ack
#> 8   25  41756          Lucia Kmecova  4.4227616 63.48330    ack
#> 9   53  39791      Najla Masoodpanah  4.2146304 67.69793    ack
#> 10   4  34964           Corey Fisher  3.7033585 71.40129    ack
#> 11  61  33676         Benjamin Gorko  3.5669346 74.96822    ack
#> 12  17  30540              Adam John  3.2347720 78.20300    ack
#> 13  22  23722     Michael Lingelbach  2.5126150 80.71561    ack
#> 14 109  20863             Alex Bates  2.2097920 82.92540    ack
#> 15  50  18729       Shahrozia Imtiaz  1.9837605 84.90916    ack
#> 16 164  16946    Mahmoud Elbahnasawi  1.7949066 86.70407    ack
#> 17 167  16306           Arian Jamasb  1.7271183 88.43119    ack
#> 18  29  16082         Nadiya Sharifi  1.7033924 90.13458    ack
#> 19  75  14785          Philipp Ranft  1.5660152 91.70060    ack
#> 20  55  10190          Greg Jefferis  1.0793165 92.77991    ack
#> 21  27   7885             Sri Murthy  0.8351728 93.61509    ack
#> 22 127   7591             Lisa Marin  0.8040326 94.41912    ack
#> 23 118   5064         Clement Hallou  0.5363748 94.95549    ack
#> 24 120   4582 Amelia Edmondson-Stait  0.4853217 95.44081    ack
#> 25  44   4581             Ala Haddad  0.4852158 95.92603    ack
#> 26 123   4134             Fiona Love  0.4378699 96.36390    ack
#> 27  28   3769           Bindu Gampah  0.3992094 96.76311    ack
#> 28  76   3540           Steven Calle  0.3749539 97.13806    ack
#> 29   2   3040              Davi Bock  0.3219943 97.46006    ack

write_authors(pnsc, auth=3.0)

#> Ruairi Roberts, Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Philipp Schlegel, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Alex Bates, Shahrozia Imtiaz, Mahmoud Elbahnasawi, Arian Jamasb, Nadiya Sharifi, Philipp Ranft, Greg Jefferis, Sri Murthy, Lisa Marin, Clement Hallou, Amelia Edmondson-Stait, Ala Haddad, Fiona Love, Bindu Gampah, Steven Calle, Davi Bock for contributing 43 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')