Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(
  skids,
  type = c("nodes", "synapses", "pre", "post"),
  ...
)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributions towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

# \donttest{
pnsc=summarise_contribution("annotation:^PN$")

#> Warning: `arrange_()` is deprecated as of dplyr 0.7.0.
#> Please use `arrange()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Warning: `mutate_()` is deprecated as of dplyr 0.7.0.
#> Please use `mutate()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

suggest_authorship(pnsc)

#> Warning: `filter_()` is deprecated as of dplyr 0.7.0.
#> Please use `filter()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#>     id      n              full_name        pct     cpct action
#> 1   94 288988       Philipp Schlegel 22.7512524 22.75125   auth
#> 2  117 161054         Ruairi Roberts 12.6793507 35.43060   auth
#> 3   12 110296          Jacob Ratliff  8.6833091 44.11391   auth
#> 4   11  80444        bailey harrison  6.3331410 50.44705   auth
#> 5   26  74574             Jawaid Ali  5.8710116 56.31806   auth
#> 6   13  62987           Addy Adesina  4.9587981 61.27686    ack
#> 7   54  42936        Ben Koppenhaver  3.3802364 64.65710    ack
#> 8   25  41752          Lucia Kmecova  3.2870233 67.94412    ack
#> 9   53  40006      Najla Masoodpanah  3.1495654 71.09369    ack
#> 10   4  34984           Corey Fisher  2.7541968 73.84788    ack
#> 11  61  33667         Benjamin Gorko  2.6505129 76.49840    ack
#> 12  17  30538              Adam John  2.4041751 78.90257    ack
#> 13  22  23715     Michael Lingelbach  1.8670185 80.76959    ack
#> 14 109  21606             Alex Bates  1.7009826 82.47057    ack
#> 15  50  18714       Shahrozia Imtiaz  1.4733032 83.94388    ack
#> 16 164  17028    Mahmoud Elbahnasawi  1.3405689 85.28445    ack
#> 17 167  16973           Arian Jamasb  1.3362389 86.62068    ack
#> 18  29  16108         Nadiya Sharifi  1.2681398 87.88882    ack
#> 19 251  15612       Nikolas Drummond  1.2290910 89.11792    ack
#> 20 127  15220             Lisa Marin  1.1982299 90.31615    ack
#> 21  75  14742          Philipp Ranft  1.1605982 91.47674    ack
#> 22  55  10237          Greg Jefferis  0.8059316 92.28268    ack
#> 23 123   8051             Fiona Love  0.6338337 92.91651    ack
#> 24  27   7877             Sri Murthy  0.6201351 93.53664    ack
#> 25 271   7261        Robert Turnbull  0.5716391 94.10828    ack
#> 26 283   6357           Imaan Tamimi  0.5004696 94.60875    ack
#> 27  44   4589             Ala Haddad  0.3612797 94.97003    ack
#> 28 120   4574 Amelia Edmondson-Stait  0.3600988 95.33013    ack
#> 29 264   4549           Levi Helmick  0.3581306 95.68826    ack
#> 30 118   4127         Clement Hallou  0.3249077 96.01317    ack
#> 31  76   3905           Steven Calle  0.3074302 96.32060    ack
#> 32  28   3769           Bindu Gampah  0.2967233 96.61732    ack
#> 33 272   3393           Billy Morris  0.2671218 96.88444    ack
#> 34   2   3320              Davi Bock  0.2613747 97.14582    ack
#> 35 245   3123         Farzaan Salman  0.2458654 97.39168    ack

write_authors(pnsc, auth=3.0)

#> Philipp Schlegel, Ruairi Roberts, Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah

write_ack(pnsc)

#> We thank Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Alex Bates, Shahrozia Imtiaz, Mahmoud Elbahnasawi, Arian Jamasb, Nadiya Sharifi, Nikolas Drummond, Lisa Marin, Philipp Ranft, Greg Jefferis, Fiona Love, Sri Murthy, Robert Turnbull, Imaan Tamimi, Ala Haddad, Amelia Edmondson-Stait, Levi Helmick, Clement Hallou, Steven Calle, Bindu Gampah, Billy Morris, Davi Bock, Farzaan Salman for contributing 41.1 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')

#> Warning: `group_by_()` is deprecated as of dplyr 0.7.0.
#> Please use `group_by()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Warning: `summarise_()` is deprecated as of dplyr 0.7.0.
#> Please use `summarise()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

# }