Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids

catmaid skeleton ids (see catmaid_skids)

type

Whether to summarise contributinons towards arbour (nodes) or synapses

...

Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x

A data frame containing the output of summarise_contribution

auth

The minimum percentage of the total effort for authorship

ackn

The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   12 110422          Jacob Ratliff 15.8146956 15.81470   auth
#> 2   11  80683        bailey harrison 11.5554607 27.37016   auth
#> 3   26  70255             Jawaid Ali 10.0619572 37.43211   auth
#> 4   13  63207           Addy Adesina  9.0525390 46.48465   auth
#> 5   54  42311        Ben Koppenhaver  6.0598032 52.54446   auth
#> 6   25  39592          Lucia Kmecova  5.6703866 58.21484   auth
#> 7   53  37898      Najla Masoodpanah  5.4277710 63.64261   auth
#> 8    4  34348           Corey Fisher  4.9193382 68.56195    ack
#> 9   61  31122         Benjamin Gorko  4.4573088 73.01926    ack
#> 10  17  28372              Adam John  4.0634524 77.08271    ack
#> 11 117  26064         Ruairi Roberts  3.7328995 80.81561    ack
#> 12  22  21093     Michael Lingelbach  3.0209503 83.83656    ack
#> 13  50  18610       Shahrozia Imtiaz  2.6653338 86.50190    ack
#> 14  29  15715         Nadiya Sharifi  2.2507104 88.75261    ack
#> 15  75  14827          Philipp Ranft  2.1235306 90.87614    ack
#> 16  55   9285          Greg Jefferis  1.3298025 92.20594    ack
#> 17  27   7848             Sri Murthy  1.1239946 93.32993    ack
#> 18 120   4888 Amelia Edmondson-Stait  0.7000619 94.03000    ack
#> 19 118   4260         Clement Hallou  0.6101194 94.64012    ack
#> 20  94   4042       Philipp Schlegel  0.5788973 95.21901    ack
#> 21  28   3769           Bindu Gampah  0.5397981 95.75881    ack
#> 22 123   3137             Fiona Love  0.4492828 96.20809    ack
#> 23  76   3052           Steven Calle  0.4371090 96.64520    ack
#> 24   2   3039              Davi Bock  0.4352471 97.08045    ack

write_authors(pnsc, auth=3.0)

#> Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Ruairi Roberts, Michael Lingelbach

write_ack(pnsc)

#> We thank Corey Fisher, Benjamin Gorko, Adam John, Ruairi Roberts, Michael Lingelbach, Shahrozia Imtiaz, Nadiya Sharifi, Philipp Ranft, Greg Jefferis, Sri Murthy, Amelia Edmondson-Stait, Clement Hallou, Philipp Schlegel, Bindu Gampah, Fiona Love, Steven Calle, Davi Bock for contributing 33.4 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')