Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria. suggest_authorship suggests how to acknowledge individuals write_authors produces a list of authors write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids)

type
Whether to summarise contributinons towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")

#> Warning: 'rbind.pages' is deprecated.
#> Use 'rbind_pages' instead.
#> See help("Deprecated")

suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   12 110390          Jacob Ratliff 14.1789588 14.17896   auth
#> 2   11  80119        bailey harrison 10.2908234 24.46978   auth
#> 3  117  74829         Ruairi Roberts  9.6113534 34.08114   auth
#> 4   26  73087             Jawaid Ali  9.3876036 43.46874   auth
#> 5   13  63201           Addy Adesina  8.1178039 51.58654   auth
#> 6   54  43030        Ben Koppenhaver  5.5269553 57.11350   auth
#> 7   25  41750          Lucia Kmecova  5.3625467 62.47605   auth
#> 8   53  39793      Najla Masoodpanah  5.1111813 67.58723   auth
#> 9    4  34859           Corey Fisher  4.4774375 72.06466    ack
#> 10  61  33943         Benjamin Gorko  4.3597826 76.42445    ack
#> 11  17  28484              Adam John  3.6586055 80.08305    ack
#> 12  22  20948     Michael Lingelbach  2.6906498 82.77370    ack
#> 13  50  18735       Shahrozia Imtiaz  2.4064027 85.18010    ack
#> 14  29  15934         Nadiya Sharifi  2.0466304 87.22673    ack
#> 15  75  14828          Philipp Ranft  1.9045711 89.13131    ack
#> 16 109  12046             Alex Bates  1.5472392 90.67855    ack
#> 17  94  11884       Philipp Schlegel  1.5264313 92.20498    ack
#> 18  55   9487          Greg Jefferis  1.2185504 93.42353    ack
#> 19  27   7890             Sri Murthy  1.0134250 94.43695    ack
#> 20  44   4625             Ala Haddad  0.5940546 95.03101    ack
#> 21 120   4609 Amelia Edmondson-Stait  0.5919995 95.62301    ack
#> 22 118   4260         Clement Hallou  0.5471724 96.17018    ack
#> 23  28   3769           Bindu Gampah  0.4841063 96.65428    ack
#> 24 123   3281             Fiona Love  0.4214255 97.07571    ack
#> 25  76   3279           Steven Calle  0.4211686 97.49688    ack
#> 26   2   3039              Davi Bock  0.3903420 97.88722    ack

write_authors(pnsc, auth=3.0)

#> Jacob Ratliff, bailey harrison, Ruairi Roberts, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Shahrozia Imtiaz, Nadiya Sharifi, Philipp Ranft, Alex Bates, Philipp Schlegel, Greg Jefferis, Sri Murthy, Ala Haddad, Amelia Edmondson-Stait, Clement Hallou, Bindu Gampah, Fiona Love, Steven Calle, Davi Bock for contributing 30.3 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')

#> Warning: 'rbind.pages' is deprecated.
#> Use 'rbind_pages' instead.
#> See help("Deprecated")