Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids

catmaid skeleton ids (see catmaid_skids)

type

Whether to summarise contributinons towards arbour (nodes) or synapses

...

Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x

A data frame containing the output of summarise_contribution

auth

The minimum percentage of the total effort for authorship

ackn

The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   12 110422          Jacob Ratliff 15.7685569 15.76856   auth
#> 2   11  80683        bailey harrison 11.5217481 27.29030   auth
#> 3   26  70493             Jawaid Ali 10.0665889 37.35689   auth
#> 4   13  63207           Addy Adesina  9.0261286 46.38302   auth
#> 5   54  42341        Ben Koppenhaver  6.0464080 52.42943   auth
#> 6   25  39614          Lucia Kmecova  5.6569851 58.08642   auth
#> 7   53  37952      Najla Masoodpanah  5.4196471 63.50606   auth
#> 8    4  34366           Corey Fisher  4.9075567 68.41362    ack
#> 9   61  31177         Benjamin Gorko  4.4521590 72.86578    ack
#> 10  17  28372              Adam John  4.0515975 76.91738    ack
#> 11 117  27326         Ruairi Roberts  3.9022259 80.81960    ack
#> 12  22  21093     Michael Lingelbach  3.0121368 83.83174    ack
#> 13  50  18611       Shahrozia Imtiaz  2.6577006 86.48944    ack
#> 14  29  15865         Nadiya Sharifi  2.2655644 88.75500    ack
#> 15  75  14828          Philipp Ranft  2.1174780 90.87248    ack
#> 16  55   9288          Greg Jefferis  1.3263512 92.19883    ack
#> 17  27   7852             Sri Murthy  1.1212866 93.32012    ack
#> 18 120   4889 Amelia Edmondson-Stait  0.6981623 94.01828    ack
#> 19 118   4260         Clement Hallou  0.6083394 94.62662    ack
#> 20  94   4053       Philipp Schlegel  0.5787792 95.20540    ack
#> 21  28   3769           Bindu Gampah  0.5382233 95.74362    ack
#> 22 123   3144             Fiona Love  0.4489716 96.19260    ack
#> 23  76   3057           Steven Calle  0.4365478 96.62914    ack
#> 24   2   3039              Davi Bock  0.4339773 97.06312    ack

write_authors(pnsc, auth=3.0)

#> Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Ruairi Roberts, Michael Lingelbach

write_ack(pnsc)

#> We thank Corey Fisher, Benjamin Gorko, Adam John, Ruairi Roberts, Michael Lingelbach, Shahrozia Imtiaz, Nadiya Sharifi, Philipp Ranft, Greg Jefferis, Sri Murthy, Amelia Edmondson-Stait, Clement Hallou, Philipp Schlegel, Bindu Gampah, Fiona Love, Steven Calle, Davi Bock for contributing 33.6 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')