Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria. suggest_authorship suggests how to acknowledge individuals write_authors produces a list of authors write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids)

type
Whether to summarise contributinons towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   12 110309          Jacob Ratliff 13.5606368 13.56064   auth
#> 2  117  98851         Ruairi Roberts 12.1520684 25.71271   auth
#> 3   11  80580        bailey harrison  9.9059561 35.61866   auth
#> 4   26  73323             Jawaid Ali  9.0138300 44.63249   auth
#> 5   13  63159           Addy Adesina  7.7643371 52.39683   auth
#> 6   54  43030        Ben Koppenhaver  5.2898150 57.68664   auth
#> 7   25  41755          Lucia Kmecova  5.1330752 62.81972   auth
#> 8   53  39794      Najla Masoodpanah  4.8920032 67.71172    ack
#> 9    4  34882           Corey Fisher  4.2881554 71.99988    ack
#> 10  61  33478         Benjamin Gorko  4.1155572 76.11543    ack
#> 11  17  28464              Adam John  3.4991702 79.61460    ack
#> 12  22  20945     Michael Lingelbach  2.5748356 82.18944    ack
#> 13  50  18735       Shahrozia Imtiaz  2.3031532 84.49259    ack
#> 14  29  15977         Nadiya Sharifi  1.9641035 86.45670    ack
#> 15 109  15664             Alex Bates  1.9256254 88.38232    ack
#> 16  75  14830          Philipp Ranft  1.8230991 90.20542    ack
#> 17  94  11975       Philipp Schlegel  1.4721249 91.67755    ack
#> 18  55   9727          Greg Jefferis  1.1957711 92.87332    ack
#> 19  27   7886             Sri Murthy  0.9694511 93.84277    ack
#> 20 164   4677    Mahmoud Elbahnasawi  0.5749585 94.41773    ack
#> 21  44   4617             Ala Haddad  0.5675825 94.98531    ack
#> 22 120   4613 Amelia Edmondson-Stait  0.5670908 95.55240    ack
#> 23 118   4260         Clement Hallou  0.5236954 96.07610    ack
#> 24  28   3769           Bindu Gampah  0.4633352 96.53943    ack
#> 25 123   3466             Fiona Love  0.4260864 96.96552    ack
#> 26  76   3418           Steven Calle  0.4201856 97.38570    ack
#> 27   2   3039              Davi Bock  0.3735940 97.75930    ack

write_authors(pnsc, auth=3.0)

#> Jacob Ratliff, Ruairi Roberts, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Shahrozia Imtiaz, Nadiya Sharifi, Alex Bates, Philipp Ranft, Philipp Schlegel, Greg Jefferis, Sri Murthy, Mahmoud Elbahnasawi, Ala Haddad, Amelia Edmondson-Stait, Clement Hallou, Bindu Gampah, Fiona Love, Steven Calle, Davi Bock for contributing 34.9 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')