Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre",
  "post"), ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributions towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

# \donttest{
pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   94 277792       Philipp Schlegel 22.2373606 22.23736   auth
#> 2  117 161001         Ruairi Roberts 12.8881944 35.12556   auth
#> 3   12 110296          Jacob Ratliff  8.8292389 43.95479   auth
#> 4   11  80444        bailey harrison  6.4395744 50.39437   auth
#> 5   26  74576             Jawaid Ali  5.9698386 56.36421   auth
#> 6   13  62987           Addy Adesina  5.0421345 61.40634   auth
#> 7   54  42938        Ben Koppenhaver  3.4372041 64.84355    ack
#> 8   25  41753          Lucia Kmecova  3.3423443 68.18589    ack
#> 9   53  40001      Najla Masoodpanah  3.2020960 71.38799    ack
#> 10   4  34984           Corey Fisher  2.8004832 74.18847    ack
#> 11  61  33667         Benjamin Gorko  2.6950568 76.88353    ack
#> 12  17  30538              Adam John  2.4445791 79.32810    ack
#> 13  22  23721     Michael Lingelbach  1.8988755 81.22698    ack
#> 14 109  21568             Alex Bates  1.7265270 82.95351    ack
#> 15  50  18714       Shahrozia Imtiaz  1.4980632 84.45157    ack
#> 16 164  17027    Mahmoud Elbahnasawi  1.3630182 85.81459    ack
#> 17 167  16972           Arian Jamasb  1.3586154 87.17320    ack
#> 18  29  16108         Nadiya Sharifi  1.2894518 88.46266    ack
#> 19 127  15089             Lisa Marin  1.2078805 89.67054    ack
#> 20  75  14742          Philipp Ranft  1.1801030 90.85064    ack
#> 21 251  12460       Nikolas Drummond  0.9974280 91.84807    ack
#> 22  55  10233          Greg Jefferis  0.8191557 92.66722    ack
#> 23 123   8012             Fiona Love  0.6413638 93.30859    ack
#> 24  27   7880             Sri Murthy  0.6307971 93.93938    ack
#> 25 271   6254        Robert Turnbull  0.5006352 94.44002    ack
#> 26  44   4591             Ala Haddad  0.3675114 94.80753    ack
#> 27 120   4574 Amelia Edmondson-Stait  0.3661505 95.17368    ack
#> 28 283   4461           Imaan Tamimi  0.3571048 95.53079    ack
#> 29 118   4127         Clement Hallou  0.3303680 95.86115    ack
#> 30  76   3985           Steven Calle  0.3190008 96.18016    ack
#> 31 264   3884           Levi Helmick  0.3109158 96.49107    ack
#> 32  28   3769           Bindu Gampah  0.3017100 96.79278    ack
#> 33 272   3395           Billy Morris  0.2717711 97.06455    ack
#> 34   2   3320              Davi Bock  0.2657673 97.33032    ack
#> 35 245   3131         Farzaan Salman  0.2506378 97.58096    ack

write_authors(pnsc, auth=3.0)

#> Philipp Schlegel, Ruairi Roberts, Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah

write_ack(pnsc)

#> We thank Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Alex Bates, Shahrozia Imtiaz, Mahmoud Elbahnasawi, Arian Jamasb, Nadiya Sharifi, Lisa Marin, Philipp Ranft, Nikolas Drummond, Greg Jefferis, Fiona Love, Sri Murthy, Robert Turnbull, Ala Haddad, Amelia Edmondson-Stait, Imaan Tamimi, Clement Hallou, Steven Calle, Levi Helmick, Bindu Gampah, Billy Morris, Davi Bock, Farzaan Salman for contributing 36.2 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')
# }