Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(
  skids,
  type = c("nodes", "synapses", "pre", "post"),
  ...
)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributions towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

# \donttest{
pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   94 288720       Philipp Schlegel 22.7871839 22.78718   auth
#> 2  117 161032         Ruairi Roberts 12.7094271 35.49661   auth
#> 3   12 110296          Jacob Ratliff  8.7050957 44.20171   auth
#> 4   11  80444        bailey harrison  6.3490310 50.55074   auth
#> 5   26  74574             Jawaid Ali  5.8857421 56.43648   auth
#> 6   13  62987           Addy Adesina  4.9712398 61.40772    ack
#> 7   54  42936        Ben Koppenhaver  3.3887175 64.79644    ack
#> 8   25  41752          Lucia Kmecova  3.2952705 68.09171    ack
#> 9   53  40006      Najla Masoodpanah  3.1574677 71.24918    ack
#> 10   4  34984           Corey Fisher  2.7611071 74.01028    ack
#> 11  61  33667         Benjamin Gorko  2.6571631 76.66745    ack
#> 12  17  30538              Adam John  2.4102072 79.07765    ack
#> 13  22  23715     Michael Lingelbach  1.8717029 80.94936    ack
#> 14 109  21580             Alex Bates  1.7031984 82.65255    ack
#> 15  50  18714       Shahrozia Imtiaz  1.4769997 84.12955    ack
#> 16 164  17028    Mahmoud Elbahnasawi  1.3439324 85.47349    ack
#> 17 167  16973           Arian Jamasb  1.3395915 86.81308    ack
#> 18  29  16108         Nadiya Sharifi  1.2713215 88.08440    ack
#> 19 127  15215             Lisa Marin  1.2008417 89.28524    ack
#> 20  75  14742          Philipp Ranft  1.1635102 90.44875    ack
#> 21 251  14647       Nikolas Drummond  1.1560123 91.60476    ack
#> 22  55  10237          Greg Jefferis  0.8079537 92.41272    ack
#> 23 123   8048             Fiona Love  0.6351872 93.04790    ack
#> 24  27   7880             Sri Murthy  0.6219279 93.66983    ack
#> 25 271   7257        Robert Turnbull  0.5727577 94.24259    ack
#> 26 283   4791           Imaan Tamimi  0.3781290 94.62072    ack
#> 27  44   4589             Ala Haddad  0.3621862 94.98290    ack
#> 28 120   4574 Amelia Edmondson-Stait  0.3610023 95.34391    ack
#> 29 264   4542           Levi Helmick  0.3584767 95.70238    ack
#> 30 118   4127         Clement Hallou  0.3257229 96.02811    ack
#> 31  76   3905           Steven Calle  0.3082016 96.33631    ack
#> 32  28   3769           Bindu Gampah  0.2974678 96.63378    ack
#> 33 272   3393           Billy Morris  0.2677920 96.90157    ack
#> 34   2   3320              Davi Bock  0.2620305 97.16360    ack
#> 35 245   3121         Farzaan Salman  0.2463245 97.40992    ack

write_authors(pnsc, auth=3.0)

#> Philipp Schlegel, Ruairi Roberts, Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah

write_ack(pnsc)

#> We thank Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Alex Bates, Shahrozia Imtiaz, Mahmoud Elbahnasawi, Arian Jamasb, Nadiya Sharifi, Lisa Marin, Philipp Ranft, Nikolas Drummond, Greg Jefferis, Fiona Love, Sri Murthy, Robert Turnbull, Imaan Tamimi, Ala Haddad, Amelia Edmondson-Stait, Levi Helmick, Clement Hallou, Steven Calle, Bindu Gampah, Billy Morris, Davi Bock, Farzaan Salman for contributing 41 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')
# }