Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributions towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1  117 150329         Ruairi Roberts 15.4639101 15.46391   auth
#> 2   12 110319          Jacob Ratliff 11.3481969 26.81211   auth
#> 3   11  80447        bailey harrison  8.2753506 35.08746   auth
#> 4   26  74657             Jawaid Ali  7.6797500 42.76721   auth
#> 5   13  62989           Addy Adesina  6.4794965 49.24670   auth
#> 6   94  51488       Philipp Schlegel  5.2964219 54.54313   auth
#> 7   54  42978        Ben Koppenhaver  4.4210227 58.96415    ack
#> 8   25  41756          Lucia Kmecova  4.2953191 63.25947    ack
#> 9   53  39811      Najla Masoodpanah  4.0952426 67.35471    ack
#> 10   4  34965           Corey Fisher  3.5967486 70.95146    ack
#> 11  61  33676         Benjamin Gorko  3.4641529 74.41561    ack
#> 12  17  30539              Adam John  3.1414587 77.55707    ack
#> 13  22  23722     Michael Lingelbach  2.4402136 79.99728    ack
#> 14 109  21090             Alex Bates  2.1694674 82.16675    ack
#> 15  50  18728       Shahrozia Imtiaz  1.9264953 84.09325    ack
#> 16 164  16894    Mahmoud Elbahnasawi  1.7378370 85.83108    ack
#> 17 167  16316           Arian Jamasb  1.6783798 87.50946    ack
#> 18  29  16137         Nadiya Sharifi  1.6599666 89.16943    ack
#> 19  75  14783          Philipp Ranft  1.5206845 90.69011    ack
#> 20 127  14041             Lisa Marin  1.4443571 92.13447    ack
#> 21  55  10190          Greg Jefferis  1.0482159 93.18269    ack
#> 22  27   7885             Sri Murthy  0.8111072 93.99380    ack
#> 23 123   6627             Fiona Love  0.6817004 94.67550    ack
#> 24 118   5064         Clement Hallou  0.5209191 95.19641    ack
#> 25  44   4591             Ala Haddad  0.4722629 95.66868    ack
#> 26 120   4582 Amelia Edmondson-Stait  0.4713371 96.14001    ack
#> 27  76   3886           Steven Calle  0.3997416 96.53976    ack
#> 28  28   3769           Bindu Gampah  0.3877061 96.92746    ack
#> 29   2   3040              Davi Bock  0.3127160 97.24018    ack

write_authors(pnsc, auth=3.0)

#> Ruairi Roberts, Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Philipp Schlegel, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Alex Bates, Shahrozia Imtiaz, Mahmoud Elbahnasawi, Arian Jamasb, Nadiya Sharifi, Philipp Ranft, Lisa Marin, Greg Jefferis, Sri Murthy, Fiona Love, Clement Hallou, Ala Haddad, Amelia Edmondson-Stait, Steven Calle, Bindu Gampah, Davi Bock for contributing 42.7 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')