Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre",
  "post"), ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributions towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1  117 155576         Ruairi Roberts 15.6261048 15.62610   auth
#> 2   12 110299          Jacob Ratliff 11.0784680 26.70457   auth
#> 3   11  80444        bailey harrison  8.0798219 34.78439   auth
#> 4   26  74616             Jawaid Ali  7.4944557 42.27885   auth
#> 5   13  62987           Addy Adesina  6.3264351 48.60529   auth
#> 6   94  56119       Philipp Schlegel  5.6366109 54.24190   auth
#> 7   54  42978        Ben Koppenhaver  4.3167245 58.55862    ack
#> 8   25  41756          Lucia Kmecova  4.1939864 62.75261    ack
#> 9   53  39897      Najla Masoodpanah  4.0072679 66.75988    ack
#> 10   4  34990           Corey Fisher  3.5144072 70.27428    ack
#> 11  61  33675         Benjamin Gorko  3.3823281 73.65661    ack
#> 12  17  30538              Adam John  3.0672468 76.72386    ack
#> 13  22  23721     Michael Lingelbach  2.3825451 79.10640    ack
#> 14 109  21372             Alex Bates  2.1466107 81.25301    ack
#> 15  50  18728       Shahrozia Imtiaz  1.8810465 83.13406    ack
#> 16 164  16871    Mahmoud Elbahnasawi  1.6945288 84.82859    ack
#> 17 167  16297           Arian Jamasb  1.6368761 86.46546    ack
#> 18  29  16138         Nadiya Sharifi  1.6209061 88.08637    ack
#> 19 127  14847             Lisa Marin  1.4912376 89.57761    ack
#> 20  75  14782          Philipp Ranft  1.4847090 91.06232    ack
#> 21  55  10192          Greg Jefferis  1.0236878 92.08601    ack
#> 22  27   7885             Sri Murthy  0.7919720 92.87798    ack
#> 23 251   7405       Nikolas Drummond  0.7437606 93.62174    ack
#> 24 123   6684             Fiona Love  0.6713432 94.29308    ack
#> 25 118   5064         Clement Hallou  0.5086298 94.80171    ack
#> 26  44   4592             Ala Haddad  0.4612220 95.26293    ack
#> 27 120   4590 Amelia Edmondson-Stait  0.4610211 95.72395    ack
#> 28  76   3985           Steven Calle  0.4002547 96.12421    ack
#> 29  28   3769           Bindu Gampah  0.3785596 96.50277    ack
#> 30   2   3320              Davi Bock  0.3334619 96.83623    ack

write_authors(pnsc, auth=3.0)

#> Ruairi Roberts, Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Philipp Schlegel, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Alex Bates, Shahrozia Imtiaz, Mahmoud Elbahnasawi, Arian Jamasb, Nadiya Sharifi, Lisa Marin, Philipp Ranft, Greg Jefferis, Sri Murthy, Nikolas Drummond, Fiona Love, Clement Hallou, Ala Haddad, Amelia Edmondson-Stait, Steven Calle, Bindu Gampah, Davi Bock for contributing 42.6 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')