Summarise tracer contributions / suggest authorship for set of neurons
Source: R/contributions.R

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributions towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1  117 148540         Ruairi Roberts 15.4062206 15.40622   auth
#> 2   12 110324          Jacob Ratliff 11.4425466 26.84877   auth
#> 3   11  80447        bailey harrison  8.3437742 35.19254   auth
#> 4   26  74615             Jawaid Ali  7.7388929 42.93143   auth
#> 5   13  62988           Addy Adesina  6.5329677 49.46440   auth
#> 6   94  49323       Philipp Schlegel  5.1156659 54.58007   auth
#> 7   54  42978        Ben Koppenhaver  4.4575774 59.03765    ack
#> 8   25  41756          Lucia Kmecova  4.3308344 63.36848    ack
#> 9   53  39793      Najla Masoodpanah  4.1272367 67.49572    ack
#> 10   4  34960           Corey Fisher  3.6259692 71.12169    ack
#> 11  61  33676         Benjamin Gorko  3.4927958 74.61448    ack
#> 12  17  30539              Adam John  3.1674335 77.78191    ack
#> 13  22  23722     Michael Lingelbach  2.4603902 80.24231    ack
#> 14 109  21048             Alex Bates  2.1830492 82.42535    ack
#> 15  50  18728       Shahrozia Imtiaz  1.9424243 84.36778    ack
#> 16 164  16921    Mahmoud Elbahnasawi  1.7550065 86.12279    ack
#> 17 167  16316           Arian Jamasb  1.6922573 87.81504    ack
#> 18  29  16126         Nadiya Sharifi  1.6725509 89.48759    ack
#> 19  75  14783          Philipp Ranft  1.5332581 91.02085    ack
#> 20 127  12081             Lisa Marin  1.2530130 92.27386    ack
#> 21  55  10190          Greg Jefferis  1.0568829 93.33075    ack
#> 22  27   7885             Sri Murthy  0.8178137 94.14856    ack
#> 23 123   6612             Fiona Love  0.6857811 94.83434    ack
#> 24 118   5064         Clement Hallou  0.5252262 95.35957    ack
#> 25  44   4591             Ala Haddad  0.4761678 95.83574    ack
#> 26 120   4582 Amelia Edmondson-Stait  0.4752343 96.31097    ack
#> 27  28   3769           Bindu Gampah  0.3909118 96.70188    ack
#> 28  76   3623           Steven Calle  0.3757691 97.07765    ack
#> 29   2   3040              Davi Bock  0.3153017 97.39295    ack

write_authors(pnsc, auth=3.0)

#> Ruairi Roberts, Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Philipp Schlegel, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Alex Bates, Shahrozia Imtiaz, Mahmoud Elbahnasawi, Arian Jamasb, Nadiya Sharifi, Philipp Ranft, Lisa Marin, Greg Jefferis, Sri Murthy, Fiona Love, Clement Hallou, Ala Haddad, Amelia Edmondson-Stait, Bindu Gampah, Steven Calle, Davi Bock for contributing 42.8 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')