Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria. suggest_authorship suggests how to acknowledge individuals write_authors produces a list of authors write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributinons towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1   12 110309          Jacob Ratliff 13.3927683 13.39277   auth
#> 2  117 104797         Ruairi Roberts 12.7235487 26.11632   auth
#> 3   11  80570        bailey harrison  9.7821151 35.89843   auth
#> 4   26  73361             Jawaid Ali  8.9068605 44.80529   auth
#> 5   13  63159           Addy Adesina  7.6682215 52.47351   auth
#> 6   54  43029        Ben Koppenhaver  5.2242104 57.69772   auth
#> 7   25  41753          Lucia Kmecova  5.0692895 62.76701   auth
#> 8   53  39796      Najla Masoodpanah  4.8316874 67.59870    ack
#> 9    4  34884           Corey Fisher  4.2353147 71.83402    ack
#> 10  61  33478         Benjamin Gorko  4.0646103 75.89863    ack
#> 11  17  28462              Adam John  3.4556108 79.35424    ack
#> 12  22  20945     Michael Lingelbach  2.5429614 81.89720    ack
#> 13  50  18735       Shahrozia Imtiaz  2.2746423 84.17184    ack
#> 14  29  16002         Nadiya Sharifi  1.9428250 86.11467    ack
#> 15 109  15678             Alex Bates  1.9034877 88.01815    ack
#> 16  75  14822          Philipp Ranft  1.7995595 89.81771    ack
#> 17  94  11990       Philipp Schlegel  1.4557225 91.27344    ack
#> 18  55   9727          Greg Jefferis  1.1809685 92.45440    ack
#> 19  27   7886             Sri Murthy  0.9574502 93.41185    ack
#> 20 164   4690    Mahmoud Elbahnasawi  0.5694194 93.98127    ack
#> 21  44   4617             Ala Haddad  0.5605564 94.54183    ack
#> 22 120   4613 Amelia Edmondson-Stait  0.5600707 95.10190    ack
#> 23 118   4262         Clement Hallou  0.5174553 95.61936    ack
#> 24  28   3769           Bindu Gampah  0.4575995 96.07696    ack
#> 25 123   3549             Fiona Love  0.4308890 96.50784    ack
#> 26 127   3530             Lisa Marin  0.4285822 96.93643    ack
#> 27  76   3421           Steven Calle  0.4153483 97.35177    ack
#> 28   2   3039              Davi Bock  0.3689692 97.72074    ack

write_authors(pnsc, auth=3.0)

#> Jacob Ratliff, Ruairi Roberts, bailey harrison, Jawaid Ali, Addy Adesina, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Shahrozia Imtiaz, Nadiya Sharifi, Alex Bates, Philipp Ranft, Philipp Schlegel, Greg Jefferis, Sri Murthy, Mahmoud Elbahnasawi, Ala Haddad, Amelia Edmondson-Stait, Clement Hallou, Bindu Gampah, Fiona Love, Lisa Marin, Steven Calle, Davi Bock for contributing 35 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')