Summarise tracer contributions / suggest authorship for set of neurons

summarise_contribution produces a data.frame of contributions by user. suggest_authorship can use this to suggest which contributors meet defined authorship/acknowledgement criteria.

suggest_authorship suggests how to acknowledge individuals

write_authors produces a list of authors

write_ack produces an acknowledgements wording.

summarise_contribution(skids, type = c("nodes", "synapses", "pre", "post"),
  ...)

suggest_authorship(x, auth = 5, ackn = c(3000, 100))

write_authors(x, ...)

write_ack(x, ...)

Arguments

skids
catmaid skeleton ids (see catmaid_skids) or a neuronlist

type
Whether to summarise contributions towards arbour (nodes) or synapses

...
Additional arguments passed by summarise_contribution to catmaid_get_contributor_stats or by write_authors/write_ack to suggest_authorship

x
A data frame containing the output of summarise_contribution

auth
The minimum percentage of the total effort for authorship

ackn
The minimum number of nodes/synapses to be acknowledged. When length of 2, the first number is assumed to be the criterion for arbour nodes, the second for synapses.

Value

A data.frame describing contributions or for write_authors/write_ack a string describing in descending order of contribution written to the console by cat.

Examples

pnsc=summarise_contribution("annotation:^PN$")
suggest_authorship(pnsc)

#>     id      n              full_name        pct     cpct action
#> 1  117 148781         Ruairi Roberts 15.4070071 15.40701   auth
#> 2   12 110324          Jacob Ratliff 11.4245949 26.83160   auth
#> 3   11  80447        bailey harrison  8.3306841 35.16229   auth
#> 4   26  74626             Jawaid Ali  7.7278908 42.89018   auth
#> 5   13  62988           Addy Adesina  6.5227184 49.41290   auth
#> 6   94  50362       Philipp Schlegel  5.2152338 54.62813   auth
#> 7   54  42978        Ben Koppenhaver  4.4505841 59.07871    ack
#> 8   25  41756          Lucia Kmecova  4.3240400 63.40275    ack
#> 9   53  39793      Najla Masoodpanah  4.1207616 67.52351    ack
#> 10   4  34961           Corey Fisher  3.6203842 71.14390    ack
#> 11  61  33676         Benjamin Gorko  3.4873161 74.63121    ack
#> 12  17  30539              Adam John  3.1624642 77.79368    ack
#> 13  22  23722     Michael Lingelbach  2.4565302 80.25021    ack
#> 14 109  21051             Alex Bates  2.1799350 82.43014    ack
#> 15  50  18728       Shahrozia Imtiaz  1.9393769 84.36952    ack
#> 16 164  16921    Mahmoud Elbahnasawi  1.7522531 86.12177    ack
#> 17 167  16316           Arian Jamasb  1.6896024 87.81138    ack
#> 18  29  16126         Nadiya Sharifi  1.6699269 89.48130    ack
#> 19  75  14783          Philipp Ranft  1.5308526 91.01216    ack
#> 20 127  12084             Lisa Marin  1.2513579 92.26351    ack
#> 21  55  10190          Greg Jefferis  1.0552248 93.31874    ack
#> 22  27   7885             Sri Murthy  0.8165307 94.13527    ack
#> 23 123   6613             Fiona Love  0.6848088 94.82008    ack
#> 24 118   5064         Clement Hallou  0.5244022 95.34448    ack
#> 25  44   4591             Ala Haddad  0.4754207 95.81990    ack
#> 26 120   4582 Amelia Edmondson-Stait  0.4744887 96.29439    ack
#> 27  28   3769           Bindu Gampah  0.3902986 96.68469    ack
#> 28  76   3624           Steven Calle  0.3752831 97.05997    ack
#> 29   2   3040              Davi Bock  0.3148070 97.37478    ack

write_authors(pnsc, auth=3.0)

#> Ruairi Roberts, Jacob Ratliff, bailey harrison, Jawaid Ali, Addy Adesina, Philipp Schlegel, Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John

write_ack(pnsc)

#> We thank Ben Koppenhaver, Lucia Kmecova, Najla Masoodpanah, Corey Fisher, Benjamin Gorko, Adam John, Michael Lingelbach, Alex Bates, Shahrozia Imtiaz, Mahmoud Elbahnasawi, Arian Jamasb, Nadiya Sharifi, Philipp Ranft, Lisa Marin, Greg Jefferis, Sri Murthy, Fiona Love, Clement Hallou, Ala Haddad, Amelia Edmondson-Stait, Bindu Gampah, Steven Calle, Davi Bock for contributing 42.7 % of reconstructed arbour cable.

pnsc=summarise_contribution("annotation:^PN$", type='synapses')